Move destruction id cache to explicit class
